+++ src/hooks/useOneDriveAuth.ts
@@ -1,10 +1,9 @@
-import { AuthenticationResult, PopupRequest } from '@azure/msal-browser'
 import { useCallback, useEffect, useState } from 'react'
 import { MicrosoftToken, MicrosoftUser } from 'microsoft'
 import usePCAInstance from './usePCAInstance'
+import useOneDriveSignIn from './useOneDriveSignIn'
 
 const GRAPH_API_ENDPOINT = 'https://graph.microsoft.com/v1.0/me'
-const TOKEN_STORAGE_KEY = 'onedrive_token'
 const GRAPH_API_FILES_ENDPOINT =
     'https://graph.microsoft.com/v1.0/me/drive/root/children'
 
@@ -15,15 +14,10 @@ interface AuthProps {
 }
 
 function useOneDriveAuth(clientId: string): AuthProps {
-    const [token, setToken] = useState<MicrosoftToken | undefined>()
     const [user, setUser] = useState<MicrosoftUser | undefined>()
     const [fileList, setFileList] = useState<any[]>([])
     const { msalInstance } = usePCAInstance(clientId)
-
-    const getStoredToken = (): MicrosoftToken | null => {
-        const storedTokenStr = localStorage.getItem(TOKEN_STORAGE_KEY)
-        return storedTokenStr ? JSON.parse(storedTokenStr) : null
-    }
+    const { token } = useOneDriveSignIn(msalInstance)
 
     const fetchFileList = useCallback(async (accessToken: string) => {
         const response = await fetch(GRAPH_API_FILES_ENDPOINT, {
@@ -53,61 +47,6 @@ function useOneDriveAuth(clientId: string): AuthProps {
 
         return response.json()
     }, [])
-
-    const handleSignIn = useCallback(async () => {
-        msalInstance && (await msalInstance.initialize())
-        const result = await signIn()
-        result && (await acquireToken())
-    }, [])
-
-    useEffect(() => {
-        const storedToken = getStoredToken()
-        if (storedToken && storedToken.expires_in > Date.now())
-            setToken(storedToken)
-        else handleSignIn()
-    }, [handleSignIn])
-
-    const signIn = async (): Promise<AuthenticationResult | null> => {
-        try {
-            const loginRequest: PopupRequest = {
-                scopes: ['user.read', 'Files.ReadWrite.All'],
-                prompt: 'select_account',
-            }
-            return await msalInstance.loginPopup(loginRequest)
-        } catch (error) {
-            console.error('Error during signIn:', error)
-            return null
-        }
-    }
-
-    const acquireToken = async () => {
-        try {
-            const accounts = msalInstance.getAllAccounts()
-            if (!accounts || accounts.length === 0) {
-                throw new Error('No accounts available. Authenticate first.')
-            }
-            const silentRequest: PopupRequest = {
-                scopes: ['user.read', 'Files.ReadWrite.All'],
-                account: accounts[0],
-            }
-            const response = await msalInstance.acquireTokenSilent(
-                silentRequest,
-            )
-            if (response.accessToken) {
-                const storeToken = {
-                    access_token: response.accessToken,
-                    expires_in: response.expiresOn!.getTime(),
-                }
-                localStorage.setItem(
-                    TOKEN_STORAGE_KEY,
-                    JSON.stringify(storeToken),
-                )
-                setToken(storeToken)
-            }
-        } catch (error) {
-            console.error('Error during token acquisition:', error)
-        }
-    }
 
     useEffect(() => {
         if (token) {
